small

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: rstumpf <rstumpf@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/08 16:38:00 by rstumpf           #+#    #+#             */
-/*   Updated: 2024/10/23 18:13:14 by rstumpf          ###   ########.fr       */
+/*   Updated: 2025/02/07 19:46:37 by rstumpf          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *string)
+int	ft_strlen(const char *string)
 {
 	int	i;
